Adding support for ThreadContext finalizer logic, and a thread refCount per context.
This will allow any thread to finalize the state of the context if it happens to be
the last thread using that context. This is preferable to trying to control the
state of the context from the "root" thread that initializes the context, since
child threads may outlive the root, and may still need the state contained in the
context. Controlling at the topmost level means potentially clearing context state
without notifying running child threads, leaving them in a broken state.