Adding concept of a thread contextualizer
ThreadContextualizer enables an application to apply custom extraction / initialization logic before a Runnable / Callable starts in a thread pool. Common actions here will be extracting a value from a ThreadLocal, then setting it in the same ThreadLocal when the Runnable/Callable starts, which will bridge the value across threads. When the Runnable/Callable completes, the value is cleared from the pooled thread's context to cleanup any memory consumption that might get lost in the pool.

This is specifically useful for Honeycomb event reporting, where the Honeycomb API binds the parent span to a ThreadLocal, and threaded calls get lost in the trace. Using a ThreadContextualizer, we're able to work around this limitation and store the parent span in our own ThreadLocal for bridging to associated thread executions.